+++ src/provider/SignatureHelpProvider/SignatureFuncUser.ts
@@ -3,6 +3,7 @@ import * as vscode from 'vscode';
 import type { CAhkFunc, TParamMetaOut } from '../../AhkSymbol/CAhkFunc';
 import { CMemo } from '../../tools/CMemo';
 import { getLStr } from '../../tools/str/removeSpecialChar';
+import type { TFnSignData } from './TFnSignData';
 
 function SignatureFuncSnipText2Sign(selectionRangeText: string): string {
     const arr0: readonly string[] = selectionRangeText.split('\n');
@@ -18,12 +19,6 @@ function SignatureFuncSnipText2Sign(selectionRangeText: string): string {
 
     return arr1.join('\n').trim();
 }
-
-type TFnSignData = {
-    SignInfo: vscode.SignatureInformation,
-    paramLength: number,
-    lastIsVariadic: boolean,
-};
 
 const UserFuncSign = new CMemo<CAhkFunc, TFnSignData>(
     (fnSymbol: CAhkFunc): TFnSignData => {
@@ -63,7 +58,6 @@ const UserFuncSign = new CMemo<CAhkFunc, TFnSignData>(
 
 export function SignUserFn(fnSymbol: CAhkFunc, comma: number): vscode.SignatureHelp | null {
     const { SignInfo, paramLength, lastIsVariadic } = UserFuncSign.up(fnSymbol);
-    // eslint-disable-next-line @typescript-eslint/prefer-optional-chain
     const commaFix: number = comma >= paramLength - 1 && lastIsVariadic
         ? paramLength - 1
         : comma;