+++ src/provider/SignatureHelpProvider/TFnSignData.ts
@@ -1,0 +1,7 @@
+import type * as vscode from 'vscode';
+
+export type TFnSignData = {
+    SignInfo: vscode.SignatureInformation,
+    paramLength: number,
+    lastIsVariadic: boolean,
+};