+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 - feat: add format option as `AhkNekoHelp.format.removeFirstCommaCommand`
 - feat: cmd - `format all file` add option to select `AhkNekoHelp.format.removeFirstCommaCommand`
 - feat: add sign of build-in function
-- fix: sign of user-def param is [variadic](https://www.autohotkey.com/docs/v1/Functions.htm#Variadic) case.
+- fix: sign of func last-param is [variadic](https://www.autohotkey.com/docs/v1/Functions.htm#Variadic) case.
 - fix: sign of nested function (4k 90-sec)<https://youtu.be/WcMzNcVWcYA>
 - fix: not show sign at func/method def-range
 - fix: core command vs Assign ...